fix: rename types/protocol
