+++ src/components/Modals/AssetTypeSelectionModal/AssetTypeSelectionModal.tsx
@@ -4,7 +4,7 @@ import _ from 'lodash'
 import { ReactComponent as CloseIcon } from 'assets/images/icon-close.svg'
 import { ModalStyles, CloseButton, ModalBody, ModalWrapper, ModalRow, ModalTitle } from 'components/Modals/styles'
 import { TypeButton } from './AssetTypeSelectionModal.styles'
-import { EAssetType } from 'types'
+import { EAssetType } from 'types/protocol'
 
 interface Props {
   open: boolean
