+++ src/components/Modals/CreatorSetupModal/CreatorSetupModal.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react'
 import * as Modal from 'react-modal'
 import { ReactComponent as CloseIcon } from 'assets/images/icon-close.svg'
 import { ModalStyles, CloseButton, ModalBody, ModalWrapper, ModalRow, ModalTitle } from 'components/Modals/styles'
-import { TEntityCreatorModel } from 'types'
+import { TEntityCreatorModel } from 'types/protocol'
 import CreatorCard from 'components/Entities/CreateEntity/CreateEntitySettings/Components/CreatorCard/CreatorCard'
 import { Button } from 'pages/CreateEntity/Components'
 import { FormData } from 'components/JsonForm/types'
