+++ src/components/Modals/LiquiditySetupModal/LiquiditySetupModal.tsx
@@ -4,7 +4,7 @@ import { ReactComponent as CloseIcon } from 'assets/images/icon-close.svg'
 import { ModalStyles, CloseButton, ModalBody, ModalWrapper, ModalRow, ModalTitle } from 'components/Modals/styles'
 import { Button } from 'pages/CreateEntity/Components'
 import { FormData } from 'components/JsonForm/types'
-import { TEntityLiquidityModel } from 'types'
+import { TEntityLiquidityModel } from 'types/protocol'
 import LiquidityCard from 'components/Entities/CreateEntity/CreateEntityAdvanced/Components/LiquidityCard/LiquidityCard'
 import { theme, Typography } from 'components/App/App.styles'
