+++ src/components/Modals/ServiceSetupModal/ServiceSetupModal.tsx
@@ -3,7 +3,7 @@ import * as Modal from 'react-modal'
 import { ReactComponent as CloseIcon } from 'assets/images/icon-close.svg'
 import { ModalStyles, CloseButton, ModalBody, ModalWrapper, ModalRow, ModalTitle } from 'components/Modals/styles'
 import { theme, Typography } from 'components/App/App.styles'
-import { TEntityServiceModel } from 'types'
+import { TEntityServiceModel } from 'types/protocol'
 import { Button } from 'pages/CreateEntity/Components'
 import { FormData } from 'components/JsonForm/types'
 import NodeCard from 'components/Entities/CreateEntity/CreateEntityAdvanced/Components/NodeCard/NodeCard'
