+++ src/redux/createEntity/createEntity.actions.ts
@@ -10,7 +10,7 @@ import {
   TEntityPaymentModel,
   TEntityServiceModel,
   TEntityTagsModel,
-} from 'types'
+} from 'types/protocol'
 import {
   ECreateEntityActions,
   TAddAssetInstancesAction,
@@ -34,9 +34,7 @@ import {
   TUpdateTagsAction,
 } from './createEntity.types'
 
-export const updateEntityTypeAction = (
-  entityType: string,
-): TUpdateEntityTypeAction => ({
+export const updateEntityTypeAction = (entityType: string): TUpdateEntityTypeAction => ({
   type: ECreateEntityActions.UpdateEntityType,
   payload: entityType,
 })
@@ -46,58 +44,42 @@ export const gotoStepAction = (no: number): TGotoStepAction => ({
   payload: no,
 })
 
-export const updateMetadataAction = (
-  metadata: TEntityMetadataModel,
-): TUpdateMetaDataAction => ({
+export const updateMetadataAction = (metadata: TEntityMetadataModel): TUpdateMetaDataAction => ({
   type: ECreateEntityActions.UpdateMetadata,
   payload: metadata,
 })
 
-export const updateCreatorAction = (
-  creator: TEntityCreatorModel,
-): TUpdateCreatorAction => ({
+export const updateCreatorAction = (creator: TEntityCreatorModel): TUpdateCreatorAction => ({
   type: ECreateEntityActions.UpdateCreator,
   payload: creator,
 })
 
-export const updateControllerAction = (
-  controller: TEntityControllerModel,
-): TUpdateControllerAction => ({
+export const updateControllerAction = (controller: TEntityControllerModel): TUpdateControllerAction => ({
   type: ECreateEntityActions.UpdateController,
   payload: controller,
 })
 
-export const updateTagsAction = (
-  tags: TEntityTagsModel,
-): TUpdateTagsAction => ({
+export const updateTagsAction = (tags: TEntityTagsModel): TUpdateTagsAction => ({
   type: ECreateEntityActions.UpdateTags,
   payload: tags,
 })
 
-export const updateServiceAction = (
-  services: TEntityServiceModel[],
-): TUpdateServiceAction => ({
+export const updateServiceAction = (services: TEntityServiceModel[]): TUpdateServiceAction => ({
   type: ECreateEntityActions.UpdateService,
   payload: services,
 })
 
-export const updatePaymentsAction = (
-  payments: TEntityPaymentModel[],
-): TUpdatePaymentsAction => ({
+export const updatePaymentsAction = (payments: TEntityPaymentModel[]): TUpdatePaymentsAction => ({
   type: ECreateEntityActions.UpdatePayments,
   payload: payments,
 })
 
-export const updateLiquidityAction = (
-  liquidity: TEntityLiquidityModel[],
-): TUpdateLiquidityAction => ({
+export const updateLiquidityAction = (liquidity: TEntityLiquidityModel[]): TUpdateLiquidityAction => ({
   type: ECreateEntityActions.UpdateLiquidity,
   payload: liquidity,
 })
 
-export const updateClaimsAction = (claims: {
-  [id: string]: TEntityClaimModel
-}): TUpdateClaimsAction => ({
+export const updateClaimsAction = (claims: { [id: string]: TEntityClaimModel }): TUpdateClaimsAction => ({
   type: ECreateEntityActions.UpdateClaims,
   payload: claims,
 })
@@ -109,31 +91,22 @@ export const updateLinkedResourceAction = (linkedResource: {
   payload: linkedResource,
 })
 
-export const updateEntityClassDidAction = (
-  did: string,
-): TUpdateEntityClassDidAction => ({
+export const updateEntityClassDidAction = (did: string): TUpdateEntityClassDidAction => ({
   type: ECreateEntityActions.UpdateEntityClassDid,
   payload: did,
 })
 
-export const updateAssetClassDidAction = (
-  did: string,
-): TUpdateAssetClassDidAction => ({
+export const updateAssetClassDidAction = (did: string): TUpdateAssetClassDidAction => ({
   type: ECreateEntityActions.UpdateAssetClassDid,
   payload: did,
 })
 
-export const addAssetInstancesAction = (
-  instances: TEntityModel[],
-): TAddAssetInstancesAction => ({
+export const addAssetInstancesAction = (instances: TEntityModel[]): TAddAssetInstancesAction => ({
   type: ECreateEntityActions.AddAssetInstances,
   payload: instances,
 })
 
-export const updateAssetInstanceAction = (
-  id: number,
-  instance: TEntityModel,
-): TUpdateAssetInstanceAction => ({
+export const updateAssetInstanceAction = (id: number, instance: TEntityModel): TUpdateAssetInstanceAction => ({
   type: ECreateEntityActions.UpdateAssetInstance,
   payload: { id, data: instance },
 })
@@ -142,16 +115,12 @@ export const removeAssetInstancesAction = (): TRemoveAssetInstancesAction => ({
   type: ECreateEntityActions.RemoveAssetInstances,
 })
 
-export const updateLocalisationAction = (
-  localisation: ELocalisation,
-): TUpdateLocalisationAction => ({
+export const updateLocalisationAction = (localisation: ELocalisation): TUpdateLocalisationAction => ({
   type: ECreateEntityActions.UpdateLocalisation,
   payload: localisation,
 })
 
-export const updatePageAction = (
-  page: TEntityPageModel,
-): TUpdatePageAction => ({
+export const updatePageAction = (page: TEntityPageModel): TUpdatePageAction => ({
   type: ECreateEntityActions.UpdatePage,
   payload: page,
 })
